+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Result can now be modified in [onError](https://superforms.rocks/concepts/events#onerror), for setting a status code for example.
+- Status codes above 400 but below 500 were not applied to the page status when the result was modified in `onUpdate`.
+
 ## [2.19.0]
 
 ### Changed